requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
anthropic>=0.40.0
litellm>=1.50.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.12.0
